v1.3.9
updates:
- support for larger integers (updated json::Scanner and json::Node to store integers in int64_t)
- improved precision for floating point numbers in json (updated json::Scanner to parse and store FP number in double)
- updated examples and tests (changes from std::get<int> to std::get<int64_t>

added:
- added warning for no exported CXX environment variable in tests/test.py script

+++ include/parser.h
@@ -3,6 +3,7 @@
 
 #include <string>
 #include <variant>
+#include <cstdint>
 
 #include "json.h"
 
@@ -57,7 +58,7 @@ private:
   [[nodiscard]]
   char peek_char() const;
 
-  std::variant<int, float, bool, std::string> value;
+  std::variant<int64_t, double, bool, std::string> value;
   const std::string& data;
   ScannerLocation    loc;
   ScannerLocation    lastloc;
